+++ lib/glimmer/swt/layout_proxy.rb
@@ -43,10 +43,11 @@ module Glimmer
       def initialize(underscored_layout_name, widget_proxy, args)
         @underscored_layout_name = underscored_layout_name
         @widget_proxy = widget_proxy
+        no_margin = args.delete(:no_margin)
         args = SWTProxy.constantify_args(args)
         @swt_layout = self.class.swt_layout_class_for(underscored_layout_name).new(*args)
-        @swt_layout.marginWidth = 15 if @swt_layout.respond_to?(:marginWidth)
-        @swt_layout.marginHeight = 15 if @swt_layout.respond_to?(:marginHeight)
+        @swt_layout.marginWidth = no_margin ? 0 : 15 if @swt_layout.respond_to?(:marginWidth)
+        @swt_layout.marginHeight = no_margin ? 0 : 15 if @swt_layout.respond_to?(:marginHeight)
         @swt_layout.marginTop = 0 if @swt_layout.respond_to?(:marginTop)
         @swt_layout.marginRight = 0 if @swt_layout.respond_to?(:marginRight)
         @swt_layout.marginBottom = 0 if @swt_layout.respond_to?(:marginBottom)