+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Add `content` method to MessageBox
 - WidgetProxy now supports taking a fully constructed swt_widget argument instead of init_args
 - Set marginLeft marginRight marginTop marginBottom to 0 when methods are available on a layout (like RowLayout)
+- Add :no_margin SWT style to layouts
 
 **CI:**